refactor(platform-wallet): extract shared candidate_address_set helper
transfer.rs::auto_select_inputs and withdrawal.rs::plan_withdrawal both built
the platform-address candidate SET with an identical block — take the
wallet-manager read lock, resolve the payment account, and union the transient
derived pool (addresses.addresses) with the hydrated address_balances map. The
block was copy-pasted (8bd3a858db retrofitted transfer onto the shape
withdrawal already had), so a future fix to how spend candidates are enumerated
would have to be made twice or the two live spend paths would silently diverge.

Extract it into PlatformAddressWallet::candidate_address_set, called from both.
Behavior is unchanged: each caller keeps its own downstream handling (transfer
tolerates an empty set; withdrawal errors early), and balances are still read
fresh from the chain via AddressInfo::fetch_many at each call site. Drops two
now-unused imports (BTreeSet, PlatformP2PKHAddress) from both callers.
